アクセストークンからユーザIDを取得してAPI処理実行する暫定コード

+++ src/main/java/com/example/backend/app/api/todo/TodoV2RestController.java
@@ -10,11 +10,13 @@ import com.example.fw.common.exception.SystemException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import java.util.Collection;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.validator.constraints.UUID;
+import org.jspecify.annotations.NonNull;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -23,33 +25,33 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import software.amazon.awssdk.services.dynamodb.model.TransactionCanceledException;
 
-/// Todoを扱うREST APIのRestControllerクラス（従来API）
+/// Todoを扱うREST APIのRestControllerクラス(OIDCによるV2 API)
 @Tag(name = "Todo", description = "Todo API")
 @XRayEnabled
 @RestController
-@RequestMapping("/api/v1/todos")
+@RequestMapping("/api/v2/todos")
 @RequiredArgsConstructor
-public class TodoRestController {
+public class TodoV2RestController {
 
     private final TodoService todoService;
     private final TodoMapper todoMapper;
 
     /// Todoリストを取得する
     ///
-    /// @param userId ユーザID
     /// @return Todoリスト
     @Operation(summary = "Todoリスト取得", description = "Todoリストを取得する。")
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public List<TodoResource> getTodos(@Parameter(name = "user_id", description = "ユーザID") //
-    @RequestParam(name = "user_id", required = false) String userId) {
-        Collection<Todo> todos = todoService.findAllByUserId(userId);
-        return todoMapper.modelsToResources(todos);
+    public List<TodoV2Resource> getTodos(@AuthenticationPrincipal Jwt jwt) {
+        // アクセストークンから preferred_username クレームを取得
+        String userId = getUserId(jwt);
+
+        var todos = todoService.findAllByUserId(userId);
+        return todoMapper.modelsToV2Resources(todos);
     }
 
     /// 指定したTodo IDに対応するTodoを取得する
@@ -59,10 +61,16 @@ public class TodoRestController {
     @Operation(summary = "Todo取得", description = "指定したTodo IDに対応するTodoを取得する。")
     @GetMapping("{todoId}")
     @ResponseStatus(HttpStatus.OK)
-    public TodoResource getTodo(
-        @Parameter(description = "Todo ID") @PathVariable @UUID String todoId) {
-        Todo todo = todoService.findOne(todoId);
-        return todoMapper.modelToResource(todo);
+    public TodoV2Resource getTodo(
+        @Parameter(description = "Todo ID") @PathVariable @UUID String todoId,
+        @AuthenticationPrincipal Jwt jwt) {
+        var userId = getUserId(jwt);
+        var todo = todoService.findOne(todoId);
+        if (userId.equals(todo.getUserId())) {
+            return todoMapper.modelTodoV2Resource(todo);
+        }
+        // TODO: ユーザIDが一致しない場合は、nullで返却する暫定仕様とする（本来は業務エラー）
+        return null;
     }
 
     /// Todoを登録する
@@ -72,11 +80,15 @@ public class TodoRestController {
     @Operation(summary = "Todo登録", description = "Todoを登録する。")
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public TodoResource postTodos(@Parameter(
-        description = "登録するTodo") @RequestBody @Validated TodoResource todoResource) {
+    public TodoV2Resource postTodos(@Parameter(
+            description = "登録するTodo") @RequestBody @Validated TodoV2Resource todoResource,
+        @AuthenticationPrincipal Jwt jwt) {
+        var userId = getUserId(jwt);
+        var todo = todoMapper.resourceToModel(todoResource);
+        todo.setUserId(userId);
         try {
-            Todo createdTodo = todoService.create(todoMapper.resourceToModel(todoResource));
-            return todoMapper.modelToResource(createdTodo);
+            var createdTodo = todoService.create(todo);
+            return todoMapper.modelTodoV2Resource(createdTodo);
         } catch (TransactionCanceledException e) {
             // 理由コードごとにハンドリングしたい場合は、以下のようにキャッチしてハンドリング
             if (DynamoDBTransactionUtil.isTransactionConditionalCheckFailed(e)) {
@@ -93,13 +105,12 @@ public class TodoRestController {
             //@formatter:off
             /*
             if (DynamoDBTransactionUtil.isTransactionConditionalCheckFailedOrConflict(e)) {
-                
+
                 throw new DynamoDBTransactionBusinessException(e, MessageIds.W_EX_5004, todoResource.getTodoTitle());
             }*/
             //@formatter:on
             throw e;
         }
-
     }
 
     /// バッチ処理向けに登録件数をチェックせずにTodoを登録する
@@ -109,11 +120,15 @@ public class TodoRestController {
     @Operation(summary = "バッチ処理用Todo登録", description = "バッチ処理向けに登録件数をチェックせずにTodoを登録する。")
     @PostMapping("batch")
     @ResponseStatus(HttpStatus.CREATED)
-    public TodoResource postTodosForBatch(@Parameter(
-        description = "登録するTodo") @RequestBody @Validated TodoResource todoResource) {
+    public TodoV2Resource postTodosForBatch(@Parameter(
+            description = "登録するTodo") @RequestBody @Validated TodoV2Resource todoResource,
+        @AuthenticationPrincipal Jwt jwt) {
+        var userId = getUserId(jwt);
+        var todo = todoMapper.resourceToModel(todoResource);
+        todo.setUserId(userId);
         try {
-            Todo createdTodo = todoService.createForBatch(todoMapper.resourceToModel(todoResource));
-            return todoMapper.modelToResource(createdTodo);
+            var createdTodo = todoService.createForBatch(todo);
+            return todoMapper.modelTodoV2Resource(createdTodo);
         } catch (TransactionCanceledException e) {
             // 理由コードごとにハンドリングしたい場合は、以下のようにキャッチしてハンドリング
             if (DynamoDBTransactionUtil.isTransactionConditionalCheckFailed(e)) {
@@ -130,7 +145,7 @@ public class TodoRestController {
             //@formatter:off
             /*
             if (DynamoDBTransactionUtil.isTransactionConditionalCheckFailedOrConflict(e)) {
-                
+
                 throw new DynamoDBTransactionBusinessException(e, MessageIds.W_EX_5004, todoResource.getTodoTitle());
             }*/
             //@formatter:on
@@ -145,11 +160,19 @@ public class TodoRestController {
     @Operation(summary = "Todo完了", description = "指定したTodo IDのTodoを完了状態に更新する。")
     @PutMapping("{todoId}")
     @ResponseStatus(HttpStatus.OK)
-    public TodoResource putTodo(
-        @Parameter(description = "Todo ID") @PathVariable @UUID String todoId) {
+    public TodoV2Resource putTodo(
+        @Parameter(description = "Todo ID") @PathVariable @UUID String todoId,
+        @AuthenticationPrincipal Jwt jwt) {
+        var userId = getUserId(jwt);
+        var todo = todoService.findOne(todoId);
+        if (!userId.equals(todo.getUserId())) {
+            // TODO: ユーザIDが一致しない場合は、nullで返却する暫定仕様とする（本来は業務エラー）
+            return null;
+        }
+        // ユーザIDが一致する場合のみ完了状態に更新する
         try {
             Todo finishedTodo = todoService.finish(todoId);
-            return todoMapper.modelToResource(finishedTodo);
+            return todoMapper.modelTodoV2Resource(finishedTodo);
         } catch (TransactionCanceledException e) {
             // 理由コードごとにハンドリングしたい場合は、以下のようにキャッチしてハンドリング
             if (DynamoDBTransactionUtil.isTransactionConditionalCheckFailed(e)) {
@@ -178,7 +201,15 @@ public class TodoRestController {
     @Operation(summary = "Todo削除", description = "指定したTodo IDのTodoを削除する。")
     @DeleteMapping("{todoId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTodo(@Parameter(description = "Todo ID") @PathVariable @UUID String todoId) {
+    public void deleteTodo(@Parameter(description = "Todo ID") @PathVariable @UUID String todoId,
+        @AuthenticationPrincipal Jwt jwt) {
+        var userId = getUserId(jwt);
+        var todo = todoService.findOne(todoId);
+        if (!userId.equals(todo.getUserId())) {
+            // TODO: ユーザIDが一致しない場合は何もしない暫定仕様とする（本来は業務エラー）
+            return;
+        }
+        // ユーザIDが一致する場合のみ削除する
         try {
             todoService.delete(todoId);
         } catch (TransactionCanceledException e) {
@@ -201,5 +232,14 @@ public class TodoRestController {
             //@formatter:on
             throw e;
         }
+
+
+    }
+
+    // TODO:ユーティリティ関数化
+    private @NonNull String getUserId(Jwt jwt) {
+        // アクセストークンから preferred_username クレームを取得
+        var userName = jwt.getClaimAsString("preferred_username");
+        return userName == null ? "" : userName;
     }
 }
